Clean Up + Print Outs Updated
Cleaned up old code/ commented code.
Print out to files now include the alphabet and letter indexes as well
as the “-“ now are replaced with “[LetterPair]” where letter pair is
the index it is at.

As of right now, I can not figure out how to make the permutations go
faster.