Optimization + File Handling Update
HUGE optimization update - With help from my professor with Big O
theory, the time was cut down from an overnight run to a little under
10 minutes.

We now delete files if they are empty.

Minor optimization implemented - skips the ends of permutations if they
are out of range and increments the current digit if we are not high
enough.