Bug Fix + Percentage Counter
Fixed the sorting of the columns - instead of ABCDEF we now sort FEDCBA

Percentage Counter will display the percent done - This is still in
development

The time to finish the program has been modified and now we don’t know
how long it will take - For perspective, one hour elapsed and still on
the first key.

Suggestion: If we can distribute 2-4 keys between several computers, we
can get lists of words a lot faster.